+++ packages/fabric/editor-common/src/schema/nodes/bodied-extension.ts
@@ -24,7 +24,8 @@ export interface Definition {
 export const bodiedExtension: NodeSpec = {
   inline: false,
   group: 'block',
-  content: 'block*',
+  content: 'block+',
+  defining: true,
   selectable: true,
   attrs: {
     extensionType: { default: '' },