+++ packages/fabric/editor-core/__tests__/plugins/media/index.ts
@@ -167,7 +167,7 @@ describe('Media plugin', () => {
           fileName: 'foo.jpg',
           fileSize: 100,
           fileMimeType: 'image/jpeg',
-          thumbnail: { width: 100, height: 100, src: '' },
+          thumbnail: { dimensions: { width: 100, height: 100 }, src: '' },
         });
 
         stateManager.updateState('bar', {
@@ -176,7 +176,7 @@ describe('Media plugin', () => {
           fileName: 'bar.png',
           fileSize: 200,
           fileMimeType: 'image/png',
-          thumbnail: { width: 200, height: 200, src: '' },
+          thumbnail: { dimensions: { width: 200, height: 200 }, src: '' },
         });
 
         expect(editorView.state.doc).toEqualDocument(