+++ packages/fabric/media-viewer/src/components/media-file-list-viewer.tsx
@@ -120,6 +120,5 @@ export class MediaFileListViewer extends Component<
     if (onClose) {
       onClose();
     }
-  }
-
+  };
 }