+++ packages/fabric/editor-test-helpers/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/editor-test-helpers
 
+## 1.8.4
+
+## 1.8.3
+- [patch] fix extension replacement with empty content [e151446](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e151446)
+
 ## 1.8.2
 
 ## 1.8.1