+++ packages/fabric/media-avatar-picker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-avatar-picker
 
+## 5.0.0
+
+## 4.6.0
+
+- [minor] [MSW-338] feat - ability to remove [9905003](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9905003)
+
 ## 4.5.1
 
 ## 4.5.0